% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dist.R
\name{calc_dist}
\alias{calc_dist}
\title{Title To calculate distance between two studies using covariate information}
\usage{
calc_dist(df, col_names, Study = "Study", Treat = "Treatment", weights, digits)
}
\arguments{
\item{df}{A data frame consists of columns namely "Study", "Treatment", and at least one covariate.}

\item{col_names}{A vector of column names specifying covariate names.}

\item{Study}{A column name in a data frame named as "Study" specifying study names.}

\item{Treat}{A column name in a data frame named as "Treatment" specifying treatment names.}

\item{weights}{A variable in which the results of specify_weight() function was stored.}

\item{digits}{A numeric value indicating the number of decimal places in the Distance calculated.}
}
\value{
Data frame
}
\description{
Title To calculate distance between two studies using covariate information
}
\examples{
attach(exampleData)
var = c("Male","Age")
weights = specify_weight(var, weights = c(0.5,0.5))
weights
dist = calc_dist(df = exampleData, col_names = var, weights = weights,digits = 4)
dist
}
\author{
Shubhram Pandey \email{shubhram1992@gmail.com}
}
