% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{Function to check if all values are numeric in data}
\usage{
check_data(df, col_names = NULL)
}
\arguments{
\item{df}{A data frame contains columns that represent covariates}

\item{col_names}{A numeric vector of covariates that can be binary or continuous}
}
\value{
logical
}
\description{
Function to check if all values are numeric in data
}
\examples{
attach(exampleData)
var = c("Age","Male")
x = check_data(df = exampleData, col_names = var)
x
}
\author{
Shubhram Pandey \email{shubhram1992@gmail.com}
}
