% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData.R
\docType{data}
\name{exampleData}
\alias{exampleData}
\title{This is a simulated data}
\format{
A data frame with with the 4 following variables (columns).
\describe{
  \item{Study}{This character vector represents number of the study.}
  \item{Male}{This vector represents the proportion of males.}
  \item{Age}{This vector represents the average age in each study.}
  \item{Treatment}{This vector represents the treatment.}
  ...
}
}
\usage{
exampleData
}
\description{
Data were extracted from the studies included.
}
\details{
A simulated data were created to run examples.
}
\author{
Shubhram Pandey \email{shubhram.pandey@heorlytics.com}
}
\keyword{datasets}
