% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_prop.R
\name{is_prop}
\alias{is_prop}
\title{Function to check if columns are proportions}
\usage{
is_prop(df, col_names)
}
\arguments{
\item{df}{a data frame to be checked}

\item{col_names}{column names to be checked}
}
\value{
list
}
\description{
Function to check if columns are proportions
}
\examples{
#' attach(exampleData)
result <- is_prop(exampleData,c("Male","Age"))
result
}
\author{
Shubhram Pandey \email{shubhram1992@gmail.com}
}
