% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_weight.R
\name{specify_weight}
\alias{specify_weight}
\title{Title specify_weight}
\usage{
specify_weight(var, weights)
}
\arguments{
\item{var}{Variables for which weights can be assigned}

\item{weights}{weights in same sequence as variables}
}
\value{
list
}
\description{
Title specify_weight
}
\examples{
var = c("Male","Age")
weights = specify_weight(var, weights = c(0.5,0.5))
weights
}
\author{
Shubhram Pandey \email{shubhram1992@gmail.com}
}
